pyproject.toml
[project]
name = "brownfield-cartographer"
version = "0.1.0"
description = "A multi-agent codebase intelligence system for rapid FDE onboarding in production environments"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-sql>=0.3.0",
    "tree-sitter-yaml>=0.6.0",
    "sqlglot>=26.0.0",
    "networkx>=3.4",
    "pydantic>=2.10.0",
    "typer>=0.15.0",
    "langgraph>=0.4.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "litellm>=1.60.0",
    "chromadb>=0.6.0",
    "scikit-learn>=1.6.0",
    "tiktoken>=0.9.0",
    "gitpython>=3.1.0",
    "rich>=13.9.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
cartographer = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
